# src/docs/swagger.yaml
definitions:
...
  models.BasicErrorRes:
    properties:
      error:
        type: string
    type: object
...
  models.Booking:
    properties:
      SVCPID:
        type: string
      bookingTimestamp:
        type: string
      cancel:
        $ref: '#/definitions/models.BookingCancel'
      endTime:
        type: string
      feedback:
        $ref: '#/definitions/models.Feedback'
      serviceID:
        type: string
      serviceName:
        type: string
      startTime:
        type: string
      status:
        $ref: '#/definitions/models.BookingStatus'
      timeslotID:
        type: string
      totalBookingPrice:
        type: number
      userID:
        type: string
    type: object
  models.BookingBasicRes:
    properties:
      message:
        type: string
      result:
        $ref: '#/definitions/models.Booking'
    type: object
  models.BookingCancel:
    properties:
      cancelBy:
        description: who cancelled
        type: string
      cancelReason:
        description: reason for cancellation
        type: string
      cancelStatus:
        description: true if cancelled
        type: boolean
      cancelTimestamp:
        description: time of cancellation
        type: string
    type: object
  models.BookingFull:
    properties:
      SVCPID:
        type: string
      SVCPName:
        type: string
      averageRating:
        type: number
      bookingID:
        type: string
      bookingTimestamp:
        type: string
      cancel:
        $ref: '#/definitions/models.BookingCancel'
      endTime:
        type: string
      feedback:
        $ref: '#/definitions/models.Feedback'
      serviceDescription:
        type: string
      serviceID:
        type: string
      serviceImg:
        items:
          type: integer
        type: array
      serviceName:
        description: service name
        type: string
      startTime:
        type: string
      status:
        $ref: '#/definitions/models.BookingStatus'
      timeslotID:
        type: string
      totalBookingPrice:
        type: number
      userID:
        type: string
    type: object
...
  models.BookingShowALL:
    properties:
      SVCPID:
        type: string
      SVCPName:
        type: string
      bookingID:
        type: string
      bookingTimestamp:
        type: string
      cancel:
        $ref: '#/definitions/models.BookingCancel'
      endTime:
        type: string
      feedback:
        $ref: '#/definitions/models.Feedback'
      serviceID:
        type: string
      serviceName:
        type: string
      startTime:
        type: string
      status:
        $ref: '#/definitions/models.BookingStatus'
      timeslotID:
        type: string
      totalBookingPrice:
        type: number
      userID:
        type: string
    type: object
  models.BookingStatus:
    properties:
      paymentStatus:
        type: boolean
      paymentTimestamp:
        type: string
      rescheduleStatus:
        description: true if rescheduled
        type: boolean
      svcpCompleted:
        type: boolean
      svcpCompletedTimestamp:
        type: string
      svcpConfirmed:
        type: boolean
      svcpConfirmedTimestamp:
        type: string
      userCompleted:
        type: boolean
      userCompletedTimestamp:
        type: string
    type: object
  models.BookingWithIdArrayRes:
    properties:
      message:
        type: string
      result:
        items:
          $ref: '#/definitions/models.BookingShowALL'
        type: array
    type: object
  models.BookkingDetailRes:
    properties:
      message:
        type: string
      result:
        $ref: '#/definitions/models.BookingFull'
    type: object
...
  models.Feedback:
    properties:
      content:
        type: string
      feedbackID:
        type: string
      rating:
        type: number
    type: object
...
  models.RequestBookingAll:
    properties:
      cancelStatus:
        type: integer
      paymentStatus:
        type: integer
      reservationType:
        type: string
      startAfter:
        type: string
      svcpCompleted:
        type: integer
      svcpConfirmed:
        type: integer
      userCompleted:
        type: integer
    type: object
  models.RequestBookingId:
    properties:
      bookingID:
        type: string
    type: object
...
  models.RequestCancelBooking:
    properties:
      bookingID:
        type: string
      cancelReason:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /service/booking/all/svcp:
    post:
      consumes:
      - application/json
      description: |-
        json body not required if you dont want to filter result
        startAfter is filter booking that has timeslot Start Before this time
        reservationType is checking booking is "incoming" or "outgoing"
        cancelStatus ,paymentStatus ,svcpConfirmed ,svcpCompleted ,userCompleted is filter booking with status 0 == false, 1 == true, 2 == dont care(or you can unuse this filed in json body)
        if dont want to filter that field dont use that field in json body
        filter is and-condition(&&)
        example {}
        example {"reservationType":"incoming","svcpCompleted": 1,"userCompleted": 0}
      parameters:
      - description: get all booking with filter(optional)
        in: body
        name: service
        schema:
          $ref: '#/definitions/models.RequestBookingAll'
      produces:
      - application/json
      responses:
        "200":
          description: get all svcp booking successfully
          schema:
            items:
              $ref: '#/definitions/models.BookingWithIdArrayRes'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: get all booking of svcp with filter(optional)
      tags:
      - Booking svcp
...
  /service/booking/cancel/svcp:
    patch:
      consumes:
      - application/json
      description: can only cancel not completed by svcp booking and not cancelled
      parameters:
      - description: booking id
        in: body
        name: bookingID
        required: true
        schema:
          $ref: '#/definitions/models.RequestCancelBooking'
      produces:
      - application/json
      responses:
        "200":
          description: Booking cancelled successfully
          schema:
            $ref: '#/definitions/models.BookingBasicRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: svcp cancel booking
      tags:
      - Booking svcp
...
  /service/booking/complete/svcp:
    patch:
      consumes:
      - application/json
      description: |-
        can only complete not completed booking by svcp ,not cancelled ,startime is pass
        if svcp not comfirmed booking it will auto comfirm first
      parameters:
      - description: booking id
        in: body
        name: bookingID
        required: true
        schema:
          $ref: '#/definitions/models.RequestBookingId'
      produces:
      - application/json
      responses:
        "200":
          description: Booking completed successfully
          schema:
            $ref: '#/definitions/models.BookingBasicRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: set complete svcp for booking by booking id
      tags:
      - Booking svcp
...
  /service/booking/confirm/svcp:
    patch:
      consumes:
      - application/json
      description: can only comfirm booking that is not comfirmed by svcp and not
        cancelled
      parameters:
      - description: booking id
        in: body
        name: bookingID
        required: true
        schema:
          $ref: '#/definitions/models.RequestBookingId'
      produces:
      - application/json
      responses:
        "200":
          description: Booking svcp completed successfully
          schema:
            $ref: '#/definitions/models.BookingBasicRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: set comfirm svcp for booking by booking id
      tags:
      - Booking svcp
...
  /service/booking/detail/svcp:
    post:
      consumes:
      - application/json
      description: get a booking detail by booking id
      parameters:
      - description: booking id
        in: body
        name: bookingID
        required: true
        schema:
          $ref: '#/definitions/models.RequestBookingId'
      produces:
      - application/json
      responses:
        "200":
          description: get detail booking
          schema:
            $ref: '#/definitions/models.BookkingDetailRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: svcp get a booking detail by booking id
      tags:
      - Booking svcp
...
swagger: "2.0"